add LevelTransition cutscene

+++ Assets/Scripts/NPC/EnemyHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    public int currentHealth;
-    public int maxHealth = 100;
+    public float currentHealth;
+    public float maxHealth = 100f;
     private List<GameObject> hitEffects = new List<GameObject>(); // Store all hit effects
 
     private void Awake()
@@ -13,7 +13,7 @@ public class EnemyHealth : MonoBehaviour
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int amount, GameObject hitEffect = null)
+    public void TakeDamage(float amount, GameObject hitEffect = null)
     {
         currentHealth -= amount;
